+++ sdk/python/tests/compiler/compiler_tests.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import kfp.compiler as compiler
 import kfp.dsl as dsl
 import os
@@ -23,7 +22,6 @@ import tarfile
 import tempfile
 import unittest
 import yaml
-import datetime
 
 class TestCompiler(unittest.TestCase):
 
@@ -141,21 +139,6 @@ class TestCompiler(unittest.TestCase):
       # Replace next line with commented line for gathering golden yaml.
       shutil.rmtree(tmpdir)
       # print(tmpdir)
-
-  def test_convert_k8s_obj_to_dic_accepts_dict(self):
-    now = datetime.datetime.now()
-    converted = compiler.Compiler()._convert_k8s_obj_to_dic({
-      "ENV": "test",
-      "number": 3,
-      "list": [1,2,3],
-      "time": now
-    })
-    self.assertEqual(converted, {
-      "ENV": "test",
-      "number": 3,
-      "list": [1,2,3],
-      "time": now.isoformat()
-    })
 
   def test_composing_workflow(self):
     """Test compiling a simple workflow, and a bigger one composed from the simple one."""
